requirements.txt
streamlit>=1.31.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
yfinance>=0.2.40,<0.3.0
scipy>=1.10.0
scikit-learn>=1.3.0
openpyxl>=3.1.0
statsmodels>=0.14.0
openai
python-dotenv
